Alterado base de dados de TXT para CSV e adicionado dados do participantes nas cartelas

+++ cartela.php
@@ -3,9 +3,6 @@ include("config.php");
 include("functions.php");
 
 $ler_cartela = ler_cartela($_GET['numero_cartela']);
-$sorteios_ranqueados = ranquear_sorteios($ler_cartela);
-
-
 $ler_todas_cartelas = ler_todas_cartelas();
 $lista_cartela_ordem = lista_cartela_ordem();
 $dezenas_chamadas = lista_dezenas();
@@ -20,33 +17,34 @@ $dezenas_chamadas = lista_dezenas();
 			</div>
 	  		<div class="col-xs-12 col-sm-9 col-md-9 col-lg-9">
 	  			<div class="cartela">
-		  			<h2>Cartela nº <?php echo link_cartela_prev_next($_GET['numero_cartela'],$ler_todas_cartelas)['prev'];?> <?php echo $_GET['numero_cartela']?> <?php echo link_cartela_prev_next($_GET['numero_cartela'],$ler_todas_cartelas)['next'];?></h2>
+		  			<h2>Cartela nº <?php echo link_cartela_prev_next($_GET['numero_cartela'],$ler_todas_cartelas)['prev'];?> <?php echo str_pad($_GET['numero_cartela'], 6, '0', STR_PAD_LEFT)?> <?php echo link_cartela_prev_next($_GET['numero_cartela'],$ler_todas_cartelas)['next'];?></h2>
 					<ul>
 						<?php 
-						foreach($ler_cartela as $dezena){	?>
-							<li class="dezena <?php echo (in_array($dezena,$dezenas_chamadas)?"sorteada":false)?>">
-								<a href="excluir_incluir_dezena.php?dezena=<?php echo $dezena; ?>&back=cartela.php&numero_cartela=<?php echo $_GET['numero_cartela']?>" title="Excluir dezena"><?php echo $dezena; ?></a>
-							</li>
+						if(is_array($ler_cartela['dezenas']) and count($ler_cartela['dezenas']) > 0 ){ ?>
+							<?php foreach($ler_cartela['dezenas'] as $dezena){	?>
+								<li class="dezena <?php echo (in_array($dezena,$dezenas_chamadas)?"sorteada":false)?>">
+									<a href="excluir_incluir_dezena.php?dezena=<?php echo $dezena; ?>&back=cartela.php&numero_cartela=<?php echo $_GET['numero_cartela']?>" title="Excluir dezena"><?php echo $dezena; ?></a>
+								</li>
+							<?php } ?>
+
+							<?php if($ler_cartela['telefone'] || $ler_cartela['telefone'] || $ler_cartela['email']){ ?>
+								<div class="titular">
+									<?php if($ler_cartela['telefone']){ ?>
+										<h3><?php echo $ler_cartela['nome']?></h3>
+									<?php } ?>
+									<?php if($ler_cartela['telefone']){ ?>
+										<span class="telefone"><?php echo $ler_cartela['telefone']?></span>
+									<?php } ?>
+
+									<?php if($ler_cartela['email']){ ?>
+										<span class="e-mail"><?php echo $ler_cartela['email']?></span>
+									<?php } ?>
+								</div>
+							<?php } ?>
 						<?php } ?>
 					</ul>
 				</div>
 
-				<div class="sorteios">
-			
-		  			<div class="cartela">
-			  			<h2>Sorteios anteriores</h2>
-						
-						<?php 
-						foreach($sorteios_ranqueados as $sorteio){ ?>
-					
-								<h3><?php echo formata_data($sorteio['data'])?> - Rodada nº <?php echo $sorteio['rodada']; ?> <?php echo $sorteio['pontos']; ?> pontos.</h3>
-									
-						
-						<?php } ?>
-						
-					</div>
-
-				</div>
 				<div class="btn_voltar">
 					<a href="./" class="btn btn-success" title="VOLTAR">VOLTAR</a>
 				</div>
